ci: update task runners and workflows to handle the monorepo structure dynamically (#13)
This PR replaces the current `tox.ini`, which is hardcoded to run the tests for the `pathops` package, with a `justfile` that can be parametrized by package name.

The `justfile` provides two public variables, `python` for setting the Python version, and `package` for specifying which package to run tests for. The monorepo's top-level `pyproject.toml` is updated to specify the development dependencies, allowing `uv run` to install these for linting and tests. To install the package itself for testing and static analysis, an optional dependency group is specified for each package, instructing `uv` to install the package from the local files.

The github `tests` workflow is updated to invoke `just` instead of `tox`, passing the appropriate arguments. It also features a new job which checks which packages have changes, and which test suites they have, so that only tests for the changed packages are run in PRs. If global files like the `justfile` or `pyproject.toml` are changed, then tests are run for all packages regardless of other changes.

Where simple commands are not possible in `tests.yaml` and the `justfile`, scripts are used, in-line bash scripts in the `justfile`, and a separate Python script in `.github` for `tests.yaml`.