tests: add Juju integrations tests (#16)
This PR adds simple Juju integration tests for pathops. To test with the
in-development version of pathops, the pack.sh script copies the pathops
library into the charm before packing.

This PR updates the justfile to have separate pack-k8s, pack-vm, juju-k8s and
juju-vm recipes. These pack the kubernetes or machine charm respectively, and
execute the integration tests for the respective substrate.